[aimedb] Change extId type to integer

+++ src/main/java/icu/samnyan/aqua/sega/chunithm/dao/userdata/UserActivityRepository.java
@@ -16,5 +16,5 @@ public interface UserActivityRepository extends JpaRepository<UserActivity, Long
 
     Optional<UserActivity> findTopByUserAndActivityIdAndKindOrderByIdDesc(UserData user, int activityId, int kind);
 
-    List<UserActivity> findAllByUser_Card_ExtIdAndKindOrderBySortNumberDesc(long extId, int kind);
+    List<UserActivity> findAllByUser_Card_ExtIdAndKindOrderBySortNumberDesc(int extId, int kind);
 }
