+++ src/main/java/icu/samnyan/aqua/sega/general/dao/CardRepository.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 @Repository("SegaCardRepository")
 public interface CardRepository extends JpaRepository<Card, Long> {
 
-    Optional<Card> findByExtId(long extId);
+    Optional<Card> findByExtId(int extId);
 
     Optional<Card> findByLuid(String luid);
 }
