+++ src/main/java/icu/samnyan/aqua/sega/general/service/CardService.java
@@ -21,6 +21,6 @@ public class CardService {
     }
 
     public Optional<Card> getCardByExtId(String extId) {
-        return cardRepository.findByExtId(Long.parseLong(extId));
+        return cardRepository.findByExtId(Integer.parseInt(extId));
     }
 }
